renamed couple test directories
