+++ tests/test_utils/utils/common/readFile.test.js
@@ -1,7 +1,9 @@
 import fs from "fs";
-import { readFile } from "../../../src/utils/common";
+import { readFile } from "../../../../src/utils/common";
+import {randomRepoPath} from "../../../helpers/helpers";
 
-const filePath = "tests/files/empty.txt";
+const repoPath = randomRepoPath();
+const filePath = `${repoPath}/empty.txt`;
 const fileData = "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. ";
 
 beforeEach(() => {
@@ -9,15 +11,13 @@ beforeEach(() => {
         fs.rmSync(filePath);
     }
     // Create directories
-    fs.mkdirSync("tests/files", { recursive: true });
+    fs.mkdirSync(repoPath, { recursive: true });
     fs.writeFileSync(filePath, fileData);
-
 });
 
 afterEach(() => {
     fs.rmSync(filePath);
 });
-
 
 test('reads file', () => {
     expect(readFile(filePath)).toBe(fileData);
